feature-crawl_all_data
Added functions to delete wanted file, file appender and crawler to crawl all raw data from pdga and append it to a file.

If the crawler for some reason fails during the crawl it can be easily continued from the failed ID. Collecting the raw data instead of parsing it during crawl makes easier to make changes to the parsing process.